refactor(app): split useCapTableWrites by action
Drop the composer. Dashboard wires useCreateStockClass,
useCreateShareholder, useIssueStockWrite, and useTransferStockWrite.
Shared wallet-ready and activity helpers live in writeHost.ts.
