+++ app/WARP.md
@@ -54,7 +54,7 @@ pnpm generate:wagmi   # Regenerate src/generated.ts from chain ABIs
   - Design system: `theme.ts`, `global-style.ts`, `typography.tsx`, `elements.tsx`, `forms.tsx`, `layout.tsx` (+ `PageHeader.tsx`)
   - `shell/` — `AppShell` (shell root), `TopBar` (system bar), `SideNav` (working nav), `AppShellContext`, `navConfig`, `WalletButtonClient` (re-export)
   - `wallet/` — native connect modal + account menu (`WalletButton`, `WalletModal`, `AccountMenu`, connector ordering helpers)
-  - `cap-table/` — `CapTableDashboard.tsx` shell (heal/reconcile/refresh/activity + wallet writes). Screens match the COMPANY left nav: `Holdings`, `StockClasses`, `Shareholders`, `IssueStock`, `Transfer`, `Transactions`. `forms/*` are fields (Issuer, Stakeholder, StockClass, IssueStock, TransferStock, MintActions, IssuerHeader, HoldingsTable), not nav destinations. Also `OwnershipBoxes`, `SetupChecklist`, `ownershipModel`, `types`, `useWalletReceipt`, `useCapTableWrites`.
+  - `cap-table/` — `CapTableDashboard.tsx` shell (heal/reconcile/refresh/activity). Screens match the COMPANY left nav: `Holdings`, `StockClasses`, `Shareholders`, `IssueStock`, `Transfer`, `Transactions`. Wallet writes: `useCreateStockClass`, `useCreateShareholder`, `useIssueStockWrite`, `useTransferStockWrite` + `useWalletReceipt`. `forms/*` are fields, not nav destinations. Also `OwnershipBoxes`, `SetupChecklist`, `ownershipModel`, `types`.
   - Shared list UI: `DataTable` + `Table` / `TableFrame` (full-width framed tables); `Modal`, `TxSuccessModal`
 - `e2e/` — Playwright specs + `mocks.ts` fixtures (`playwright.config.ts` at app root)
 - `src/hooks/` — `useMintIssuer`, `useDirectCreateStockClass`, `useDirectCreateStakeholder`, `useDirectIssueStock`, **`useDirectTransferStock`**, `useOnchainAction`, `useResource`, `useCapTableManager`
