+++ memory/sep-12-2026.md
@@ -1,0 +1,13 @@
+# Writes-hook split — left off 2026-09-12
+
+`CapTableDashboard` is 384 lines after the nav-named screens (#285). The leftover blob was `useCapTableWrites` (547).
+
+This change:
+
+- Deleted the composer. The dashboard wires four action hooks: `useCreateStockClass`, `useCreateShareholder`, `useIssueStockWrite`, `useTransferStockWrite`.
+- Shared receipt helper stays `useWalletReceipt`; shared wallet-ready / activity helpers live in `writeHost.ts`.
+- No product behavior change. `data-testid` values unchanged.
+
+Still true from `memory/sep-7-2026.md`: keep `/create`, `acceptStock`, `SAMPLE_ISSUER`, landing impl links; no beacon upgrade; do not hand-edit `generated.ts`.
+
+Next: `acceptStock` UI is future work.
