+++ memory/sep-7-2026.md
@@ -27,4 +27,4 @@ Do not:
 - Delete `/create` or `acceptStock`.
 
 Agent docs: `WARP.md`, `app/WARP.md`, `AGENTS.md`.
-Next: merge #285 → #286 → #287; `acceptStock` UI is future work.
+#285–#287 merged to `main`. Writes-hook split: `memory/sep-12-2026.md`.
